created a base abstract class to eliminate code duplication
