+++ src/WebJobs.Script.WebHost/ContainerManagement/LinuxContainerInitializationHostService.cs
@@ -18,12 +18,12 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.ContainerManagement
     public class LinuxContainerInitializationHostService : IHostedService
     {
         private readonly IEnvironment _environment;
-        private readonly IInstanceManager _instanceManager;
+        private readonly LinuxInstanceManager _instanceManager;
         private readonly ILogger _logger;
         private readonly StartupContextProvider _startupContextProvider;
         private CancellationToken _cancellationToken;
 
-        public LinuxContainerInitializationHostService(IEnvironment environment, IInstanceManager instanceManager, ILogger<LinuxContainerInitializationHostService> logger, StartupContextProvider startupContextProvider)
+        public LinuxContainerInitializationHostService(IEnvironment environment, LinuxInstanceManager instanceManager, ILogger<LinuxContainerInitializationHostService> logger, StartupContextProvider startupContextProvider)
         {
             _environment = environment;
             _instanceManager = instanceManager;
