+++ src/WebJobs.Script.WebHost/Management/InstanceManager.cs
@@ -19,10 +19,9 @@ using Newtonsoft.Json;
 
 namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
 {
-    public class InstanceManager : IInstanceManager
+    public class InstanceManager : LinuxInstanceManager
     {
         private static readonly object _assignmentLock = new object();
-        private static HostAssignmentContext _assignmentContext;
 
         private readonly ILogger _logger;
         private readonly IMetricsLogger _metricsLogger;
@@ -36,7 +35,8 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
 
         public InstanceManager(IOptionsFactory<ScriptApplicationHostOptions> optionsFactory, IHttpClientFactory httpClientFactory, IScriptWebHostEnvironment webHostEnvironment,
             IEnvironment environment, ILogger<InstanceManager> logger, IMetricsLogger metricsLogger, IMeshServiceClient meshServiceClient, IRunFromPackageHandler runFromPackageHandler,
-            IPackageDownloadHandler packageDownloadHandler)
+            IPackageDownloadHandler packageDownloadHandler) : base(httpClientFactory, webHostEnvironment,
+            environment, logger, metricsLogger, meshServiceClient)
         {
             _client = httpClientFactory?.CreateClient() ?? throw new ArgumentNullException(nameof(httpClientFactory));
             _webHostEnvironment = webHostEnvironment ?? throw new ArgumentNullException(nameof(webHostEnvironment));
@@ -49,7 +49,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
             _optionsFactory = optionsFactory ?? throw new ArgumentNullException(nameof(optionsFactory));
         }
 
-        public async Task<string> SpecializeMSISidecar(HostAssignmentContext context)
+        public override async Task<string> SpecializeMSISidecar(HostAssignmentContext context)
         {
             // No cold start optimization needed for side car scenarios
             if (context.IsWarmupRequest)
@@ -103,65 +103,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
             return null;
         }
 
-        public bool StartAssignment(HostAssignmentContext context)
-        {
-            if (!_webHostEnvironment.InStandbyMode)
-            {
-                // This is only true when specializing pinned containers.
-                if (!context.Environment.TryGetValue(EnvironmentSettingNames.ContainerStartContext, out string startContext))
-                {
-                    _logger.LogError("Assign called while host is not in placeholder mode and start context is not present.");
-                    return false;
-                }
-            }
-
-            if (_environment.IsContainerReady())
-            {
-                _logger.LogError("Assign called while container is marked as specialized.");
-                return false;
-            }
-
-            if (context.IsWarmupRequest)
-            {
-                // Based on profiling download code jit-ing holds up cold start.
-                // Pre-jit to avoid paying the cost later.
-                Task.Run(async () => await _packageDownloadHandler.Download(context.GetRunFromPkgContext()));
-                return true;
-            }
-            else if (_assignmentContext == null)
-            {
-                lock (_assignmentLock)
-                {
-                    if (_assignmentContext != null)
-                    {
-                        return _assignmentContext.Equals(context);
-                    }
-                    _assignmentContext = context;
-                }
-
-                _logger.LogInformation($"Starting Assignment. Cloud Name: {_environment.GetCloudName()}");
-
-                // set a flag which will cause any incoming http requests to buffer
-                // until specialization is complete
-                // the host is guaranteed not to receive any requests until AFTER assign
-                // has been initiated, so setting this flag here is sufficient to ensure
-                // that any subsequent incoming requests while the assign is in progress
-                // will be delayed until complete
-                _webHostEnvironment.DelayRequests();
-
-                // start the specialization process in the background
-                Task.Run(async () => await Assign(context));
-
-                return true;
-            }
-            else
-            {
-                // No lock needed here since _assignmentContext is not null when we are here
-                return _assignmentContext.Equals(context);
-            }
-        }
-
-        public async Task<string> ValidateContext(HostAssignmentContext assignmentContext)
+        public override async Task<string> ValidateContext(HostAssignmentContext assignmentContext)
         {
             _logger.LogInformation($"Validating host assignment context (SiteId: {assignmentContext.SiteId}, SiteName: '{assignmentContext.SiteName}'. IsWarmup: '{assignmentContext.IsWarmupRequest}')");
             RunFromPackageContext pkgContext = assignmentContext.GetRunFromPkgContext();
@@ -214,7 +156,12 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
             }
         }
 
-        private async Task<(string Error, long? ContentLength)> ValidateBlobPackageContext(RunFromPackageContext context)
+        protected override async Task<string> DownloadWarmupAsync(RunFromPackageContext context)
+        {
+            return await _packageDownloadHandler.Download(context);
+        }
+
+        private async Task<(string, long?)> ValidateBlobPackageContext(RunFromPackageContext context)
         {
             string blobUri = context.Url;
             string eventName = context.IsWarmUpRequest
@@ -277,38 +224,8 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
             }
         }
 
-        private async Task Assign(HostAssignmentContext assignmentContext)
+        protected override async Task ApplyContext(HostAssignmentContext assignmentContext)
         {
-            try
-            {
-                // first make all environment and file system changes required for
-                // the host to be specialized
-                await ApplyContext(assignmentContext);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Assign failed");
-                await _meshServiceClient.NotifyHealthEvent(ContainerHealthEventType.Fatal, GetType(), "Assign failed");
-                throw;
-            }
-            finally
-            {
-                // all assignment settings/files have been applied so we can flip
-                // the switch now on specialization
-                // even if there are failures applying context above, we want to
-                // leave placeholder mode
-                _logger.LogInformation("Triggering specialization");
-                _webHostEnvironment.FlagAsSpecializedAndReady();
-
-                _webHostEnvironment.ResumeRequests();
-            }
-        }
-
-        private async Task ApplyContext(HostAssignmentContext assignmentContext)
-        {
-            _logger.LogInformation($"Applying {assignmentContext.Environment.Count} app setting(s)");
-            assignmentContext.ApplyAppSettings(_environment, _logger);
-
             // We need to get the non-PlaceholderMode script Path so we can unzip to the correct location.
             // This asks the factory to skip the PlaceholderMode check when configuring options.
             var options = _optionsFactory.Create(ScriptApplicationHostOptionsSetup.SkipPlaceholder);
@@ -446,21 +363,6 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
                 _logger.LogError(e, $"Failed to mount BYOS storage account {storageInfoValue.Id}");
                 return false;
             }
-        }
-
-        public IDictionary<string, string> GetInstanceInfo()
-        {
-            return new Dictionary<string, string>
-            {
-                { "FUNCTIONS_EXTENSION_VERSION", ScriptHost.Version },
-                { "WEBSITE_NODE_DEFAULT_VERSION", "8.5.0" }
-            };
-        }
-
-        // for testing
-        internal static void Reset()
-        {
-            _assignmentContext = null;
         }
     }
 }
