+++ src/WebJobs.Script.WebHost/Management/LegionInstanceManager.cs
@@ -8,140 +8,40 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs.Script.Diagnostics;
+using Microsoft.Azure.WebJobs.Script.WebHost.Management.LinuxSpecialization;
 using Microsoft.Azure.WebJobs.Script.WebHost.Models;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
 
 namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
 {
-    public class LegionInstanceManager : IInstanceManager
+    public class LegionInstanceManager : LinuxInstanceManager
     {
-        private static readonly object _assignmentLock = new object();
-        private static HostAssignmentContext _assignmentContext;
-
-        private readonly ILogger _logger;
-        private readonly IMetricsLogger _metricsLogger;
-        private readonly IMeshServiceClient _meshServiceClient;
-        private readonly IEnvironment _environment;
-        private readonly HttpClient _client;
-        private readonly IScriptWebHostEnvironment _webHostEnvironment;
-
         public LegionInstanceManager(IHttpClientFactory httpClientFactory, IScriptWebHostEnvironment webHostEnvironment,
-            IEnvironment environment, ILogger<LegionInstanceManager> logger, IMetricsLogger metricsLogger, IMeshServiceClient meshServiceClient)
-        {
-            _client = httpClientFactory?.CreateClient() ?? throw new ArgumentNullException(nameof(httpClientFactory));
-            _webHostEnvironment = webHostEnvironment ?? throw new ArgumentNullException(nameof(webHostEnvironment));
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-            _metricsLogger = metricsLogger ?? throw new ArgumentNullException(nameof(metricsLogger));
-            _meshServiceClient = meshServiceClient ?? throw new ArgumentNullException(nameof(meshServiceClient));
-            _environment = environment ?? throw new ArgumentNullException(nameof(environment));
-        }
+            IEnvironment environment, ILogger<LegionInstanceManager> logger, IMetricsLogger metricsLogger, IMeshServiceClient meshServiceClient) : base(httpClientFactory, webHostEnvironment,
+            environment, logger, metricsLogger, meshServiceClient) { }
 
-        public Task<string> SpecializeMSISidecar(HostAssignmentContext context)
+        public override Task<string> SpecializeMSISidecar(HostAssignmentContext context)
         {
             // Legion will take care of MSI Specialization
             return Task.FromResult((string)null);
         }
 
-        public bool StartAssignment(HostAssignmentContext context)
+        protected override Task<string> DownloadWarmupAsync(RunFromPackageContext context)
         {
-            if (!_webHostEnvironment.InStandbyMode && !context.Environment.TryGetValue(EnvironmentSettingNames.ContainerStartContext, out string startContext))
-            {
-                // This is only true when specializing pinned containers.
-                _logger.LogError("Assign called while host is not in placeholder mode and start context is not present.");
-                return false;
-            }
-
-            if (_environment.IsContainerReady())
-            {
-                _logger.LogError("Assign called while container is marked as specialized.");
-                return false;
-            }
-
-            if (context.IsWarmupRequest)
-            {
-                // No need to download since file already exists
-                return true;
-            }
-
-            if (_assignmentContext != null)
-            {
-                // No lock needed here since _assignmentContext is not null when we are here
-                return _assignmentContext.Equals(context);
-            }
-
-            lock (_assignmentLock)
-            {
-                if (_assignmentContext != null)
-                {
-                    return _assignmentContext.Equals(context);
-                }
-                _assignmentContext = context;
-            }
-
-            _logger.LogInformation("Starting Assignment. Cloud Name: {cloudName}", _environment.GetCloudName());
-
-            // set a flag which will cause any incoming http requests to buffer
-            // until specialization is complete
-            // the host is guaranteed not to receive any requests until AFTER assign
-            // has been initiated, so setting this flag here is sufficient to ensure
-            // that any subsequent incoming requests while the assign is in progress
-            // will be delayed until complete
-            _webHostEnvironment.DelayRequests();
-
-            // start the specialization process in the background
-            //Task.Run(async () => await Assign(context));
-            Task.Run(() => Assign(context));
-
-            return true;
+            return Task.FromResult((string)null);
         }
 
-        public Task<string> ValidateContext(HostAssignmentContext assignmentContext)
+        public override Task<string> ValidateContext(HostAssignmentContext assignmentContext)
         {
             // Don't need to validate RunFromPackageContext in Legion
             return Task.FromResult((string)null);
         }
 
-        private async Task Assign(HostAssignmentContext assignmentContext)
+        protected override Task ApplyContext(HostAssignmentContext assignmentContext)
         {
-            try
-            {
-                // first make all environment and file system changes required for
-                // the host to be specialized
-                _logger.LogInformation("Applying {environmentCount} app setting(s)", assignmentContext.Environment.Count);
-                assignmentContext.ApplyAppSettings(_environment, _logger);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Assign failed");
-                await _meshServiceClient.NotifyHealthEvent(ContainerHealthEventType.Fatal, GetType(), "Assign failed");
-                throw;
-            }
-            finally
-            {
-                // all assignment settings/files have been applied so we can flip
-                // the switch now on specialization
-                // even if there are failures applying context above, we want to
-                // leave placeholder mode
-                _logger.LogInformation("Triggering specialization");
-                _webHostEnvironment.FlagAsSpecializedAndReady();
-                _webHostEnvironment.ResumeRequests();
-            }
-        }
-
-        public IDictionary<string, string> GetInstanceInfo()
-        {
-            return new Dictionary<string, string>
-            {
-                { EnvironmentSettingNames.FunctionsExtensionVersion, ScriptHost.Version },
-                { EnvironmentSettingNames.WebsiteNodeDefaultVersion, "8.5.0" }
-            };
-        }
-
-        // for testing
-        internal static void Reset()
-        {
-            _assignmentContext = null;
+            return null;
         }
     }
 }
